bundle_firmware: allow blobs to defer their processing
In order to make it easier for blobs to be generated from
other other blobs (which are also generated) provide mechanism
for a blob to defer its processing. That way the ordering of the
nodes in the device tree doesn't matter nor is it required to
ensure the ordering is hard coded within the code itself.

BUG=chrome-os-partner:44827
BUG=chromium:445938
BRANCH=None
TEST=Built and ran on glados.

Change-Id: I0dcfa5f4a67044839a65591d5c6522537c2576c3
Signed-off-by: Aaron Durbin <adurbin@chromiium.org>
Reviewed-on: https://chromium-review.googlesource.com/319320
Commit-Ready: Aaron Durbin <adurbin@chromium.org>
Tested-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>