bundle_firmware: ensure correct RO CBFS source file for RW CBFS blobs
The 'coreboot' blob type within the fmap.dts specifies the location
of the RO CBFS. Additionally, the cbfs_files passed to the bundling
script specify additional files to add to each CBFS. In order for
the RO CBFS to be in sync with each RW CBFS defer the blob
processing of RW CBFS regions until the RO CBFS region has been
processed. That way the cbfs_files specified are duplicated to the
RW CBFS regions.

BUG=chrome-os-partner:44827
BUG=chromium:445938
BRANCH=None
TEST=Built for chell with RW CBFS regions. Verified CBFS files
     match between regions.

Change-Id: Ie9e89294e3c5fd66ae01db8a1de947bf4a0de8da
Signed-off-by: Aaron Durbin <adurbin@chromiium.org>
Reviewed-on: https://chromium-review.googlesource.com/320560
Commit-Ready: Aaron Durbin <adurbin@chromium.org>
Tested-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>