dev: let 'timeout' exception raised.
The stateful update process won't raise errors since rootfs update can be
adopted to continue auto-update. However, a timer is set for the whole
auto-update process in case that this process is hung. This timeout exception
may be caught by stateful update and won't be raised.

This CL is used to raise 'timeout' exception in auto-update, to make sure that
the auto-update process terminates itself when it meets the deadline.

BUG=chromium:633280
TEST=Locally ran 'ds.auto_update([dut], [image_path])';
Ran 'repair' for dut from local autotest instance;
Ran unittest, devserver_integration_test.

Changes to be committed:
       modified:   cros_update.py

Change-Id: Ib52cd9204b8152a83011a6b2f13fc16656fdb60c
Reviewed-on: https://chromium-review.googlesource.com/364622
Commit-Ready: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>