cros_bundle_firmware: Add test for image equivalence
Since the goal is to create an image that's equivalent to what the
packer is generating, test for that. Fail hard iff there's a difference.

One exception is the fmap region: Both versions are generated from the
same source (ultimately chromeos.fmd in the coreboot tree), but they
differ in the order of the fmap entries and in the fmap minor version
(1.0 vs 1.1).
Another is RW_SHARED, which is never used as-is (as it's usually an
inner node in the fmap region tree)

BUG=chromium:595715
BRANCH=none
TEST=It now passes the test, but failed with some earlier commits (that
refer to this one) not applied.

Change-Id: I205487dd7c102d24606ce3657597394bc92818de
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/397219
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>