cros_bundle_firmware: reuse existing GBB
Instead of creating a new GBB, reuse the one put in by the coreboot
ebuild. We still write the flags and hwid, but the pubkeys aren't
updated anymore.
A minor change in output is that the GBB region is filled up with 0x00
instead of 0xff.

BUG=chromium:595715
BRANCH=none
TEST=bit-identical result except as explained above.

Change-Id: Idce6195da885f242494cd31990e288d7ae1190d0
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/415327
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>