devserver: Remove confusing log message.
If we are unable to import android_build, we either don't expect to use
it (e.g. on a DUT) in which case this error message just adds to
confusion in the logs, or we'll use it and blow up because the module is
set to None.

BUG=chromium:651520
TEST=None.

Change-Id: I266baee1526165af8487ffb257a0bf3eff02d3c9
Reviewed-on: https://chromium-review.googlesource.com/427998
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>