cros_update.py: do not fail provision on missed packet
This change prevents the loss of a single ping packet from causing
a provisioning failure.  The ping check is useful for speeding
up failures in the interactive case ("cros flash" and "cros
deploy" for instance) but it can be damaging in the test lab, where
packet losses can be more common and need to be tolerated better.

BUG=chromium:683304
TEST=none

Change-Id: Iff32074d469c53654e73024481060bc4e4b8f773
Reviewed-on: https://chromium-review.googlesource.com/433370
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>