Force Launch Control API to return enough results for artifact lookup.
The value of maxResults used to be defaulted 0 to include all artifacts.
The API was changed so that 0 means only 10 results are returned.
This change forces the API to return 1000 results assume that's large
enough to include all artifacts for an Android build.

BUG=b:62141068
TEST=local devserver run

Change-Id: Ib0ba83d04ea8d8d7c78b886835e25ebc845f8fe2
Reviewed-on: https://chromium-review.googlesource.com/517275
Reviewed-by: Kris Rambish <krisr@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>
Tested-by: Kris Rambish <krisr@chromium.org>
Trybot-Ready: Kris Rambish <krisr@chromium.org>