[devserver] Allow handling of clobber_stateful variable.
chromite/lib/autoupdater.py already has a clobber_stateful variable but
the devserver does not pass it to the constructor. We just leave it at the
default of True. autoupdate_EndToEndTest needs this to be False in certain
conditions so we will pass it in.

The if statement here allows this CL to be submitted now without waiting for
any other CLs to catch up and will not break provision_AutoUpdate.

BUG=chromium:709710
TEST=autoupdate_EndToEnd with this CL

Change-Id: Iac5aa7ca1349f282dd4ed1e31128f5dddd5bb30e
Reviewed-on: https://chromium-review.googlesource.com/554167
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>