Make sure chromite is properly imported at devserver startup.
At startup, before importing any modules from chromite, it's
necessary to import the 'chromite' package proper, in order to force
chromite/__init__.py to run.  Because Python.

This adjusts the import order in devserver.py to honor the
requirement, and adds a "here be dragons" comment for the unwary.

BUG=chromium:757003
TEST=start the devserver, see it still work

Change-Id: I5f1d00c76a71fb985db4e1b11a65ed6461b1d149
Reviewed-on: https://chromium-review.googlesource.com/621670
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Paul Hobbs <phobbs@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>