key increment test: accomodate unibuild shellballs
Add model to the interface and read the path from setvars.sh in the model
specific dir when a model is specified.  Changes to the further downstream
//src/platform/crostestutils/provingground/firmware/keyinc_setup.sh
to use the new model interface will be considered separately (the
functionality of the old interface without model is maintained).

BUG=b:66966755
TEST=ran fm_and_key_version_test_prep against a unibuild fw and test
image and non-unibuild ones:

./fm_and_key_version_test_prep.sh -b coral -i \
/tmp/chromiumos_test_image.bin -f 10068.12.0 -s \
/tmp/image.bin -v Google_Coral.10068.12.0 -m robo360

./fm_and_key_version_test_prep.sh -b kefka -i \
/tmp/chromiumos_test_image_kefka.bin -f 7287.379.0 -s /tmp/image_kefka.bin -v \
Google_Kefka.7287.379.0

Change-Id: I1f1c5bb2045c7d121182d13c6f20b1dd4d5c6f6a
Reviewed-on: https://chromium-review.googlesource.com/801251
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>