autoupdate: Return response with client's app ID
Currently in autoupdate we sent a default hard coded app ID to the
client. However, this is not the bahvior of omaha, as omaha always, reply with
the same app ID requested by the user. This CL, modifes this bahavior in
autoupdate to return the app ID associated with the client's request.

BUG=chromium:821514
TEST=cros flash

Change-Id: I7cd36c8e7ed6e36bed257a5373e62483136b585e
Reviewed-on: https://chromium-review.googlesource.com/961585
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>