fromupstream: Don't replace tag with "BACKPORT:" if it wasn't "UPSTREAM:"
Recent changes to fromupstream would automatically change the tag to
"BACKPORT:" if there were any conflicts when applying.  The problem is
that "BACKPORT:" implies a backport of a patch that was "UPSTREAM:".
Specifically we wouldn't use just "BACKPORT:" for a "FROMGIT:" or a
"FROMLIST:" commit.

Historically we've just skipped the backport tag when we had FROMGIT
or FROMLIST, but perhaps better is to append the two.  Thus, after
this commit:

  UPSTREAM: - clean pick from Linux
  FROMGIT: - clean pick from some other git tree
  FROMLIST: - clean pick from patchwork
  BACKPORT: - pick from Linux w/ conflicts
  BACKPORT: FROMGIT: - pick from some other git tree w/ conflicts
  BACKPORT: FROMLIST: - pick from patchwork w/ conflicts

BUG=None
TEST=Pick something with conflicts.

Change-Id: I1f9c1c432fcf134b33315050f8633f990cdb5261
Reviewed-on: https://chromium-review.googlesource.com/1035456
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>