GS Cache: Bind to either network port or socket, not both.
Before this change, gs_archive_server cannot just bind to a socket because
there is always a default port to listen. This makes it hard to start multiple
instances to bind on sockets (because we have to assign ports for all of
instanceses even when we don't want to).

In this change, I changed the option --socket and --port to be a mutual
exclusive group, which means one and only one could be specified. This
makes it easier for Puppet to start multiple intances.

BUG=chromium:824580
TEST=Ran and verified the sever with different combination of options.

Change-Id: I9f72b51cdcb694ca1f3bce0a039f3c038fc41629
Reviewed-on: https://chromium-review.googlesource.com/1067709
Commit-Ready: Congbin Guo <guocb@chromium.org>
Tested-by: Congbin Guo <guocb@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>