fromupstream: drop patchwork URL support
Providing a full URL for patchwork is inconsistent with the fromgit://
and linux:// syntax, and even within patchwork, it isn't really
flexible; it only support patchwork.kernel.org, and pwclient doesn't
really have a nice way to support arbitrary URLs like this anyway. If we
want to extend patchwork support to support other projects/servers, we
should settle on a different syntax.

BUG=none
TEST=fromupstream --help, fromupstream ... pw://XXX still work

Change-Id: I097f032c606fa2331c936a028814bc53a6f995e8
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1082925
Reviewed-by: Douglas Anderson <dianders@chromium.org>