fromupstream.py: Do not assume that Change-Id: exists
It is possible to create a commit without Change-Id:, for example by
cherry-picking a patch directly from a non-Gerrit repository. If such
a patch is encountered and --replace is specified, the script bailed
out with an error message. Check if a Change-Id exists before using it
to fix the problem.

Change-Id: Idc402123aa9f2484049a2fb8ebb118e4fc40aba2
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1152466
Reviewed-by: Brian Norris <briannorris@chromium.org>