cros_choose_profile: drop dead code
Nothing in here uses toolchain_utils.sh, and we've deleted the file,
so stop trying to source it.

The --board_overlay option hasn't been used since CL:262922 in 2015.

The --brick option has been dead for a while and the last caller was
dropped in CL:1144120.

The /etc/make.profile symlink was last used in 2015 and upgraded away
from in CL:272054.

BUG=chromium:530677
TEST=precq passes

Change-Id: I4c8e9712a48ad4c5f1c6a449c48d46e0d10cb81c
Reviewed-on: https://chromium-review.googlesource.com/1173692
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>