roafteriniter: Add a cache implementation
This commit adds in an implementation of file backed memory sharing
across processes. This allows for the kernel to be built with -j<NRPROCS>,
where each process loads in an instance of the plugin, and these plugins
are able to have a shared cache.

The following files are added in:
- cache.h: Cache implementation
- cache_test.cc: Unit tests
- cachereader.c: Utility to read a cache

BUG=chromium:860068
TEST=cache_test.cc

Change-Id: I50b9cbedcca3c6921d75829040e5d9af30ac4ec7
Reviewed-on: https://chromium-review.googlesource.com/1244402
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Commit-Ready: Zubin Mithra <zsm@chromium.org>
Tested-by: Zubin Mithra <zsm@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>