nebraska: Parse XML requests
XML-formatted update and install requests are now parsed and logged. On
receiving a request, the server parses each app request and determines
its type based on its content and whether there is an update request
associated with the platform appid. Each request is logged and the
server responds with an error.

BUG=chromium:892450
TEST=Manual, unittesting

Change-Id: Iec81c5e01a0d296bf399d5f6e5d67c7db8a25c23
Reviewed-on: https://chromium-review.googlesource.com/1281245
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Colin Howes <chowes@google.com>
Reviewed-by: Amin Hassani <ahassani@chromium.org>