fromupstream: Allow picking from git repos directly
It's sort of annoying to have to add remotes for one-off git trees of
maintainers to pick into the tree. Add a way to pick patches from a git
tree directly as long as the form follows git://repo#branch/sha1. To
differentiate this from the 'fromgit' syntax, let's deprecate
fromgit://remote/branch/sha1 and replace it with
git://remote/branch/sha1 by updating the regex to look for lines that
don't have the '#' symbol.

This makes it a little faster to pick patches from random git repos.

BUG=None
TEST=Pick with new syntax and see if it works

Change-Id: Ia31743bc56800bef69fb54360e95f387bb87bd9e
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1308315
Reviewed-by: Douglas Anderson <dianders@chromium.org>