fromupstream: Add --buganizer and --crbug options
The BUG= line is sometimes hard to get right if you forget a b: and that
bug is in buganizer you'll most likely make an impossible link. Let's
add --crbug and --buganizer to be more explicit about the bug types and
then do the 'b:' and 'chromium:' prefix for the user. They can just
specify the right command line (possibly many times) and get a list of
bugs in their BUG line.

BUG=None
TEST=Use --crbug, --buganizer, --bug, all, none, and see if all match
expectations

Change-Id: Ibaf46f8d1878b9a802703d83cd004bdb3d87cd96
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1308314
Reviewed-by: Douglas Anderson <dianders@chromium.org>