nebraska: Detect update/install operation.
An install request is distinguished from an update request based on
whether or not the platform app includes an update_check tag. Currently,
we don't have any reliable way of detecting whether an appid is for the
platform app, since update engine does not distinguish between DLCs and
the platform app when constructing a request. Since update requests are
never mixed with events or pings, we make the assumption that if the
update_check tag is omitted from any app request, it is the platform app
and the request is for an install.

BUG=chromium:892450
TEST=Unittests still pass, manual testing.

Change-Id: I14cfe82395fdaf6e9f1e7853e077facf73d48dcf
Reviewed-on: https://chromium-review.googlesource.com/1336774
Commit-Ready: Colin Howes <chowes@google.com>
Tested-by: Colin Howes <chowes@google.com>
Reviewed-by: Amin Hassani <ahassani@chromium.org>