stateful_update: Increase the timeout stateful update timeout
Daisy-stake devices are failing because it takes more than two minutes to unzip
the update.tgz and write it to the stateful (It takes 2m51 on a DUT with the
payload in /tmp). Increase this timeout so we don't hit this problem again if
the size of the stateful update grows or the disk is slow.

Also added check for the exit code of the call that does the actual work so we
know if it fails for any other reason.

BUG=chromium:912705
TEST=deployed to a DUT and stateful_update passed successfully.

Change-Id: Ib7479f5b46be83fd0c2bdc062e42c211b1229c44
Reviewed-on: https://chromium-review.googlesource.com/1370869
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>