dev-install: drop support for dynamic package.provided
We should consider the set of provided packages to be a fixed attribute
of the rootfs which means it's defined once at build_image time.  Trying
to add more packages to that list doesn't make sense from the dev_install
perspective.  Plus, we've never used this functionality.  Punt it.

BUG=chromium:202477
TEST=running dev_install on a base image still works

Change-Id: I82796075bc614121c0900518dc0724001e513529
Reviewed-on: https://chromium-review.googlesource.com/1492773
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>