nebraska: Break --payload-addr into two flags
Currently, the nebraska adds /update or /install to the --payload-addr
flag based on whether the request is install or update. But this is
unfair to the server/service that serves the payload. They have to
change their code semantics in order to handle this. Even more
complicated we send an Omaha request to a devserver address like:
"http://devserver-address/update" so the devserver so can send a
response back. Now adding /update or /install to the end of addresses
like this can become more confusing and cumbersome.

So instead of nebraska making such assumptions, we should directly pass
the addresses for install and updates to nebraska. If no address for
install was passed we use the update address instead.

So flag --payload-addr was replaced by two flags
--update-payloads-address and --install-payloads-address.

BUG=chromium:920404
TEST=unittest
TEST=./nebraska.py

Change-Id: Iaebc3e78a4d80f7085464c87b62622d7129ea66b
Reviewed-on: https://chromium-review.googlesource.com/1518803
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>