nebraska: Deprecate the version handling
We originally added a version handling to detect the version mismatches
in full and delta payloads. But truly that is not really necessary. We
normally generate payloads for a specific version. But sometimes the
version number is not really obvious or we can't catch it easily. It is
just as fine to ignore the version checking. IMO the version checking is
a call for headache for a few reasons:

- We can't really check always for the uprevs in version number because
  we will be testing enterprise rollback feature, and there we always
  use a downreved version. So there is not reason to check if an
  incoming request has lower/equal version than what the payloads
  metadata says.

- In some instances we can't even correctly extract the payloads'
  version. For example for DLC images, since they are build at
  build_package time, the version number is non-existent in the image
  itself. We might be able to somehow get the version number in the
  release builders, but locally that assumption doesn't really hold.

- Version numbers are quite complicated and versions build locally have
  weird assignment that can be complicated to parse and compare, etc.

We have other means of protecting against bad versions and nebraska
should not be the place for that. For example, if we send the wrong DLC
payload, the dlcservice will complain later. But for its worth,
currently, the devserver does not do any version checking anyway. It
always responds with version 99999.0.0.0. Update engine also doesn't
really care about the version anymore. Besides, originally we decided to
add these version so we can serve multiple different versioned payloads
in a directory and let nebraska choose one. But that is a bad idea. Not
only it introduces the randomness that is caused by nebraska decisions,
it shouldn't be the nebraska's business making those decisions. Also, we
probably would've never have this case anyway. We always update/install
a specific version. Serving different versions in one directory doesn't
really make sense.

So deprecate anything related to
that.

BUG=chromium:920404
TEST=unittest
TEST=manually running nebraska and query it.

Change-Id: Ifbb70f8e5484abdeed1856f847068e90d3ae3ae0
Reviewed-on: https://chromium-review.googlesource.com/1592554
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>