nebraska: Allow no_update option for response
Sometimes devserver might have the payloads objects to pass to the
client, but it might be overloaded and can't do that. So we need a way
to let nebraska to respond with status="noupdate"" even if we had the
payloads. This CL adds a new parameter to GetResponseTorequest to make
sure the correct flags are passed.

BUG=chromium:920404
TEST=unittest
TEST=nebraska.py

Change-Id: I69c9e1f0cbbeafbdfd5c77aa41f4b0333f3bbf0e
Reviewed-on: https://chromium-review.googlesource.com/1640973
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>