nebraska: Check more error conditions on the request
There are a few attributes that are kind of mandatory on at least on app
tag in the list of request apps. Make sure they exist and there are not
different values for them in different apps of the same request.

Also restructure some unittests a little bit with default values so the
code is cleaner and more readable.

BUG=chromium:920404
TEST=unittest
TEST=nebraska.py

Change-Id: Iae56b89abe7e1d6c70da1b49b6ca1b54fbeb51f5
Reviewed-on: https://chromium-review.googlesource.com/1640967
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>