nebraska: Handle critical updates
We need to be able to handle critical updates for testing this feature
in autotests. Critical updates (or blocking updates) are defined when
the response to the client contains the attribute deadline="now" in one
of the actions tag along with other parameters passed to the client.

Users of Nebraska as a library can pass the critical_update argument to
GetResponsetorequest() for this feature.

BUG=chromium:920404
TEST=unittest
TEST=nebraska.py

Change-Id: Ie33653f7c36e260f3c972ae09707dd527e8fe126
Reviewed-on: https://chromium-review.googlesource.com/1640969
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>