nebraska: Include public key in response if requested
Update engine client accepts a public key in the response to verify the
payload. We also use this feature to autotest the payload verification
in the client with custom signed payloads. This CL allows reading the
public key from the payload's json file and passing it to the client
when requested.

BUG=chromium:920404
TEST=unittest
TEST=nebraska.py

Change-Id: I9fc662fcd3bae20ed4f1384fef4bc65b9934248a
Reviewed-on: https://chromium-review.googlesource.com/1640970
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>