nebraska: Allow AppData with partial or empty AppID
The current problem is when we generate the payload properties from the
payload itself, we don't have access to the appid of the payload. So the
appid value would be empty. However, we still need to match based on
appid. But for now we can relax the requirement of the matching by
allowing appid match partially. Partial means if the AppData had an
empty or partial appid, we sub-string check it against the request appid
and if it was a sub-string, we consider that a match. This would work
for many of our current situations specially for the platform updates
where there is only one payload so matching with an empty appid should
work.

BUG=chromium:920404
TEST=unittest
TEST=send a request with a valid APP ID to a nebraska that has AppData
with empty APP ID and response was correct. Same for an AppData with a
substring of the request's APP ID.

Change-Id: I35e83a78c81c5d29f2d32a2e2d0febaec03b3249
Reviewed-on: https://chromium-review.googlesource.com/1674247
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>