nebraska: Allow passing critical_update=true in the URL
We need a way to respond to an update/install request with
'deadline=now' (means it is a critical update). This is mostly for
tests and cros flash because the updater does not do update if the OOBE
is not processed yet. So in most cases we need to request a critical
update so the AU does not block it.

This CL allows calling into nebraska server with URLs like:
http://<ip>:<port>/critical_update=true so we can respond with
'deadline=now'

BUG=chromium:920404
TEST= sudo ./nebraska.py --port 10000;
Then this for non critical updates:
  curl -X POST -d @request.xml http://localhost:10000;
And this for critical updates:
  curl -X POST -d @request.xml http://localhost:10000/critical_update=true;
TEST=platform/dev/nebraska/run_unittests

Change-Id: Ic83783150865061da8a9d99f94595c1fa9c11f10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1708892
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>