merge-kernel: Add support for reverts to apply as part of merge
In some cases, it may be necessary to revert one or more patches as part
of a merge. This can happen, for example, if bug fixes in the merged
branches conflict with each other, or if an API is changed differently
in both branches. This feature should only be used if reverts can not be
handled as separate commits.

BUG=b:183621345
TEST=Use script

Change-Id: Id31cb999b844a6007f3541d9df24d1530e1bec65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1733909
Reviewed-by: William K Lin <wklin@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>