start_devserver: Exec the devserver instead of backgrounding it
Starting devserver in the background and using 'wait' causes the exit
code of devserver to be dropped because the 'wait' builtin will always
return 0 when called without arguments, except on a signal.  Handling
the signal and exit code is fragile. Since the script does no more work,
exec the devserver instead.

Additionally, exec the devserver directly instead of through python, to
help with python2->3 migration. And drop the existence check which was
added to help move the devserver script.

BUG=chromium:995577
TEST='./start_devserver; echo $?' and kill with ctrl-c. exit status is 0
TEST='./start_devserver; echo $?' and kill with ctrl-\. exit status is 131
TEST=start_devserver in another shell and then './start_devserver; echo $?'
     exit status is 70 because another server is listening on the port
TEST=introduce a syntax error into devserver.py and './start_devserver; echo $?'
     exit status is 1 because the python interpreter found a syntax
     error
TEST='./start_devserver; echo $?' and kill -9 devserver.py. exit status
     is 137 (128 + signal 9)

Change-Id: I0fb9cd575c73870e59b0a74bb90f8c6a4238c816
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/1761989
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>
Tested-by: Anand Mistry <amistry@chromium.org>
Commit-Queue: Anand Mistry <amistry@chromium.org>